refactor phr types and services
