+++ abdm/urls.py
@@ -35,7 +35,10 @@ router.register(
 router.register("v3/health_id", HealthIdViewSet, basename="abdm__v3__health_id")
 router.register("v3/hip", HIPViewSet, basename="abdm__v3__hip")
 router.register("v3/hiu", HIUViewSet, basename="abdm__v3__hiu")
+
+## PHR Routes
 router.register("v3/phr/health_id", PhrAuthViewSet, basename="abdm__v3__phr_health_id")
+
 
 ## Utility Routes
 router.register(
